performance is acceptable, beta version

+++ lib/CodeGen/RegAllocRL.h
@@ -121,7 +121,7 @@ protected:
 
   /// Method called when the allocator is about to remove a LiveInterval.
   virtual void aboutToRemoveInterval(LiveInterval &LI) {}
-  void observe(std::vector<float> old_state, unsigned action, float reward, std::vector<float> new_state);
+  void observe(std::vector<int> old_state, unsigned action, float reward, std::vector<int> new_state);
 
 public:
   /// VerifyEnabled - True when -verify-regalloc is given.
@@ -131,7 +131,7 @@ public:
   static bool inference;
   static long int _score;
   static float prev_weight;
-  static std::vector<float> _state;
+  static std::vector<int> _state;
   static float curr_weight;
   SmallVector<unsigned, 8> past_cand;
