+++ lib/CodeGen/rl.h
@@ -17,10 +17,10 @@ class RandomPolicy {
 
 class QTable {
   public:
-    float get(std::vector<float> state, int action);
-    void set(std::vector<float> state, int action, float value);
-    std::pair<int, float> best(std::vector<float> state, SmallVector<unsigned, 8>& cand);
-    std::map<std::pair<std::vector<float>, int>, float> _table;
+    float get(std::vector<int> state, int action);
+    void set(std::vector<int> state, int action, float value);
+    std::pair<int, float> best(std::vector<int> state, SmallVector<unsigned, 8>& cand);
+    std::map<std::pair<std::vector<int>, int>, float> _table;
 };
 
 class GreedyQ {
@@ -29,7 +29,7 @@ class GreedyQ {
   GreedyQ(QTable* q) {
     _q = q;
   }
-  int pick_action(std::vector<float> state, SmallVector<unsigned, 8>& cand);
+  int pick_action(std::vector<int> state, SmallVector<unsigned, 8>& cand);
 
   private:
     QTable* _q;
@@ -44,7 +44,7 @@ class EpsilonPolicy {
     _random = policy_b;
     _epsilon = epsilon;
   }
-  int pick_action(std::vector<float> state, SmallVector<unsigned, 8>& cand);
+  int pick_action(std::vector<int> state, SmallVector<unsigned, 8>& cand);
 
   private:
     GreedyQ* _greedy;
@@ -62,7 +62,7 @@ class QLearner {
       _gamma = discount_rate;
     }
 
-    void observe(std::vector<float>, int action, float reward, std::vector<float>, SmallVector<unsigned, 8>& cand);
+    void observe(std::vector<int>, int action, float reward, std::vector<int>, SmallVector<unsigned, 8>& cand);
   private:
    float _alpha;
    float _gamma;
